refactor(cloud)!: Use 'image' as default vs. 'img'
